Queue file changes in the auto-archiver
This was surprisingly hard to do. Because nodejs caches imports, we need to somehow clear the import cache.
This can be done naively (https://ar.al/2021/02/22/cache-busting-in-node.js-dynamic-esm-imports/#cache-invalidation-in-esm-with-dynamic-imports) for plain js files, but it leaks memory and fails for typescript files, since the lodaer doesn’t support it.
So, we run each archiving task in a new worker thread that has its own import context. But, since the adapters we are importing are written in typescript, we need to register a loader in the worker context, because it doesn’t have it loaded by default (like the main thread does via `tsx`).
Luckly, this feature has just been added to node, but we need to bump our node version to use it. Now we can dynamically import the all chnages to modules.